Updated button clicker test; Made half working quiz maker test.
Button clicker test now goes through all activities.

Made a quiz maker test which almost works. It is able to edit text but
it will not edit the answer test and fails when it tries (for no good
reason). It does edit the title and the question fields.

Started a delete test.